django/app/requirements.txt
Django==3.0.14
environs==7.3.1
Pillow==8.2.0
djangorestframework==3.12.4